==> 2_Data_Expansion.ipynb <==
import boto3
import io
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

# export to S-3
import awswrangler as wr

# Imputation
from sklearn.impute import SimpleImputer


plt.style.use('fivethirtyeight')

import warnings
warnings.filterwarnings("ignore")

#!pip install awswrangler

pd.set_option('display.max_rows', 200)

### Load a consolidated file from S-3

session = boto3.Session()
s3 = session.client('s3')

bucket_name = 'ppsg-collections'
file_name = '2021-06-01/consolidated/random_data/consolidated_sample.parquet'

%%time
obj = s3.get_object(Bucket = bucket_name, Key = file_name)
df = pd.read_parquet(io.BytesIO(obj['Body'].read())) 

df = df.fillna(value=np.nan)

### Set Processing periods

start_date = '2020-02-01'
end_date = '2021-04-01'

#### Choose data periods

df= df[(df['nextpaymentduedt']>=start_date) & (df['nextpaymentduedt']<=end_date)]

# remove 2020 April data
df =  df[df['nextpaymentduedt'] != '2020-04-01']

### Import Functions

from important_functions import processing_data

### Processing Data (Rollback  --> Expand -- > Add one more row to handle Sunday)

%%time
df = processing_data (df,method = 'add row')

%%time
import awswrangler as wr
wr.s3.to_parquet(
    df=df,
    path='s3://ppsg-collections/2021-06-01/consolidated/random_data/consolidated_sample_expanded.parquet'
)

#df.shape

#df.head()
